refactor : Game의 책임 분리를 위한 Players 객체를 생성
